refactor : move @interactive to filter_decorator #35
