+++ test/test_decorator.py
@@ -1,5 +1,5 @@
 from interactive_pipe.headless.control import Control
-from interactive_pipe import interactive
+from interactive_pipe.helper.filter_decorator import interactive
 import numpy as np
 
 @interactive(
